Rename module to resolve naming confusion
