+++ src/content_types/simple_content.rs
