+++ src/lib.rs
@@ -1,10 +1,10 @@
-pub mod content;
+pub mod content_types;
 pub mod cyclic_graph;
 pub mod error;
 pub mod id_generator;
 pub mod node;
 
-pub use content::content::Content;
+pub use content_types::content::Content;
 pub use cyclic_graph::CyclicGraph;
 pub use error::CyclicGraphError as Error;
 pub use id_generator::{GeneratorMode, IdGenerator};
